requirements.txt
mcp>=1.0.0
python-dotenv>=1.0.0
starlette>=0.37.0
uvicorn>=0.23.0